go home

+++ format/ft_pfarg_nbr.c
@@ -6,7 +6,7 @@
 /*   By: qloubier <qloubier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/18 19:56:26 by qloubier          #+#    #+#             */
-/*   Updated: 2016/09/30 01:27:20 by qloubier         ###   ########.fr       */
+/*   Updated: 2016/10/04 19:59:15 by qloubier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,5 +15,5 @@
 void			ft_pfarg_nbr(va_list ap, t_pfc *arg)
 {
 	arg->type = PFT_INT;
-	arg->arg = va_arg(ap, int);
+	arg->arg = (intmax_t)va_arg(ap, int);
 }
